+++ servicetalk-traffic-resilience-http/src/test/java/io/servicetalk/traffic/resilience/http/CapacityClientServerTest.java
@@ -79,7 +79,7 @@ class CapacityClientServerTest {
         if (!applyOnClient) {
             TrafficResilienceHttpServiceFilter serviceFilter =
                     new TrafficResilienceHttpServiceFilter.Builder(limiterSupplier)
-                            .onRejectionPolicy(new ServiceRejectionPolicy.Builder()
+                            .rejectionPolicy(new ServiceRejectionPolicy.Builder()
                                     .onLimitResponseBuilder(ServiceRejectionPolicy.serviceUnavailable()).build())
                             .build();
