+++ servicetalk-utils-internal/src/main/java/io/servicetalk/utils/internal/NumberUtils.java
@@ -98,68 +98,4 @@ public final class NumberUtils {
         }
         return value;
     }
-
-    /**
-     * Ensures the float is greater than the min specified.
-     *
-     * @param value the float value to validate
-     * @param min the float min to validate against
-     * @param field name of the variable
-     * @return the passed value if all checks pass
-     * @throws IllegalArgumentException if the passed float doesn't meet the requirements
-     */
-    public static float ensureGreaterThan(final float value, final float min, final String field) {
-        if (value <= min) {
-            throw new IllegalArgumentException(field + ": " + value + " (expected: > " + min + ")");
-        }
-        return value;
-    }
-
-    /**
-     * Ensures the float is between 0 and 1 (inclusive).
-     *
-     * @param value the float value to validate
-     * @param field name of the variable
-     * @return the passed value if all checks pass
-     * @throws IllegalArgumentException if the passed float doesn't meet the requirements
-     */
-    public static float ensureBetweenZeroAndOne(final float value, final String field) {
-        if (value < 0.0f || value > 1.0f) {
-            throw new IllegalArgumentException(field + ": " + value + " (expected: 0.0f <= " + field + " <= 1.0f)");
-        }
-        return value;
-    }
-
-    /**
-     * Ensures the int is between the provided range (inclusive).
-     *
-     * @param value the int value to validate
-     * @param min the min int value to validate against (inclusive)
-     * @param max the max int value to validate against (inclusive)
-     * @param field name of the variable
-     * @return the passed value if all checks pass
-     * @throws IllegalArgumentException if the passed int doesn't meet the requirements
-     */
-    public static int ensureRange(final int value, final int min, final int max, final String field) {
-        if (value < min || value > max) {
-            throw new IllegalArgumentException(field + ": " + value +
-                    " (expected: " + min + " <= " + field + " <= " + max + ")");
-        }
-        return value;
-    }
-
-    /**
-     * Ensures the float is between 0 and 1 (exclusive).
-     *
-     * @param value the float value to validate
-     * @param field name of the variable
-     * @return the passed value if all checks pass
-     * @throws IllegalArgumentException if the passed float doesn't meet the requirements
-     */
-    public static float ensureBetweenZeroAndOneExclusive(final float value, final String field) {
-        if (value <= 0.0f || value >= 1.0f) {
-            throw new IllegalArgumentException(field + ": " + value + " (expected: 0.0f < " + field + " < 1.0f)");
-        }
-        return value;
-    }
 }
